+++ packages/server/api/src/app/openops-tables/template-tables/create-base-template-tables.ts
